DMT-28 Refactor main.js and donutCHart.js
Moved the render function to the donutChart class, in order to decouple the corresponding logical units. Deleted unused functions from main.js
