+++ src/main.js
@@ -4,8 +4,8 @@
  * in the license file that is distributed with this file.
  */
 
-//@ts-check - Get type warnings from the TypeScript language server. Remove if not wanted.
-import { donutChart } from "./donutChart";
+import { render } from "./donutChart";
+
 /**
  * Get access to the Spotfire Mod API by providing a callback to the initialize method.
  * @param {Spotfire.Mod} mod - mod api
@@ -13,10 +13,6 @@ import { donutChart } from "./donutChart";
 const Spotfire = window.Spotfire;
 
 Spotfire.initialize(async (mod) => {
-    /**
-     * Store the context.
-     */
-    const context = mod.getRenderContext();
 
     /**
      * Create the read function.
@@ -26,127 +22,7 @@ Spotfire.initialize(async (mod) => {
     /**
      * Initiate the read loop
      */
-    reader.subscribe(render);
-
-    /**
-     * Render the visualization
-     * @param {Spotfire.DataView} dataView
-     * @param {Spotfire.Size} size
-     */
-    async function render(dataView, size) {
-        /**
-         * Check for any errors.
-         */
-        let errors = await dataView.getErrors();
-        if (errors.length > 0) {
-            mod.controls.errorOverlay.show(errors, "dataView");
-            // TODO clear DOM
-            return;
-        }
-
-        mod.controls.errorOverlay.hide("dataView");
-
-        // Get the leaf nodes for the x hierarchy. We will iterate over them to
-        // render the bars.
-        let colorHierarchy = await dataView.hierarchy("Color");
-        let colorRoot = await colorHierarchy.root();
-        if (colorRoot == null) {
-            // Return and wait for next call to render when reading data was aborted.
-            // Last rendered data view is still valid from a users perspective since
-            // a document modification was made during a progress indication.
-            return;
-        }
-
-        let dataViewYAxis = await dataView.continuousAxis("Y");
-        if (dataViewYAxis == null) {
-            mod.controls.errorOverlay.show("No data on y axis.", "y");
-            return;
-        } else {
-            mod.controls.errorOverlay.hide("y");
-        }
-
-        // Hide tooltip
-        mod.controls.tooltip.hide();
-
-        let colorLeaves = colorRoot.leaves();
-
-        let data = colorLeaves.map((leaf) => {
-            let rows = leaf.rows();
-            return {
-                color: rows.length ? rows[0].color().hexCode : "transparent",
-                value: sumValue(rows, "Y"),
-                id: leaf.key,
-                mark: () => leaf.mark(),
-                tooltip: () => {
-                    return leaf.formattedValue() + " " + sumValue(rows, "Y");
-                }
-            };
-        });
-
-        donutChart(size, data, mod);
-
-        context.signalRenderComplete();
-    }
+    reader.subscribe(() => {
+        render(mod);
+    });
 });
-
-/**
- * Calculate the maximum value from a hierarchy. If split bars is enabled, the single maximum value from all rows will be used.
- * @param {Spotfire.DataViewHierarchyNode[]} xLeaves
- * @param {Spotfire.ModProperty<boolean>} splitBars
- * @param {number} categoricalColorCount
- */
-function calculateMaxYValue(xLeaves, splitBars, categoricalColorCount) {
-    let maxYValue = 0;
-    if (splitBars.value() && categoricalColorCount > 0) {
-        xLeaves.forEach((node) => {
-            maxYValue = Math.max(maxValue(node.rows(), "Y"), maxYValue);
-        });
-    } else {
-        xLeaves.forEach((node) => {
-            let sum = sumValue(node.rows(), "Y");
-            maxYValue = Math.max(maxYValue, sum);
-        });
-    }
-
-    return maxYValue;
-}
-
-/**
- * Calculate the total value for an axis from a set of rows. Null values are treated a 0.
- * @param {Spotfire.DataViewRow[]} rows Rows to calculate the total value from
- * @param {string} axis Name of Axis to use to calculate the value.
- */
-function sumValue(rows, axis) {
-    return rows.reduce((p, c) => +c.continuous(axis).value() + p, 0);
-}
-
-/**
- * Calculate the max value for an axis from a set of rows. Null values are treated a 0.
- * @param {Spotfire.DataViewRow[]} rows Rows to calculate the max value from
- * @param {string} axis Name of Axis to use to calculate the value.
- */
-function maxValue(rows, axis) {
-    return rows.reduce((p, c) => Math.max(+c.continuous(axis).value(), p), 0);
-}
-
-/** @returns {HTMLElement} */
-function findElem(selector) {
-    return document.querySelector(selector);
-}
-
-/**
- * Create a div element.
- * @param {string} className class name of the div element.
- * @param {string | HTMLElement} [content] Content inside the div
- */
-function createDiv(className, content) {
-    let elem = document.createElement("div");
-    elem.classList.add(className);
-    if (typeof content === "string") {
-        elem.appendChild(document.createTextNode(content));
-    } else if (content) {
-        elem.appendChild(content);
-    }
-
-    return elem;
-}
